# src/Glimmr/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Glimmr Web API",
    "description": "A simple example ASP.NET Core Web API",
    "contact": {
      "name": "d8ahazard",
      "url": "https://facebook.com/GlimmrTV",
      "email": "donate.to.digitalhigh@gmail.com"
    },
    "license": {
      "name": "GPL3.0",
      "url": "https://github.com/d8ahazard/glimmr/blob/master/COPYING"
    },
    "version": "v1.3"
  },
  "paths": {
...
    "/api/Glimmr/ambientColor": {
      "get": {
        "tags": [
          "Glimmr"
        ],
        "summary": "Retrieve the currently set ambient color.\r\nWill still return a value if ambient mode is not\r\nactive.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Glimmr"
        ],
        "summary": "Set the current ambient mode, and set device mode to ambient.",
        "requestBody": {
          "description": "A HTML-formatted RGB Color (#666666/FFFFFF)",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "type": "string"
              }
            },
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/Glimmr/ambientScene": {
      "get": {
        "tags": [
          "Glimmr"
        ],
        "summary": "Retrieve the ambient scene specified in the query, or current if none is specified.\r\nWill still return a value if ambient mode is not\r\nactive.",
        "parameters": [
          {
            "name": "sceneId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AmbientScene"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AmbientScene"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AmbientScene"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Glimmr"
        ],
        "summary": "Set the current ambient mode, and set device mode to ambient.",
        "requestBody": {
          "description": "The ID of the target ambient mode.",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "application/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "text/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AmbientScene"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AmbientScene"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AmbientScene"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Glimmr/audioScene": {
      "get": {
        "tags": [
          "Glimmr"
        ],
        "summary": "Get the currently selected audio scene.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AudioScene"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioScene"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioScene"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Glimmr"
        ],
        "summary": "Set the desired audio scene and set device mode to Audio.",
        "requestBody": {
          "description": "ID of the target audio scene.",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "application/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "text/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AudioScene"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioScene"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioScene"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Glimmr/authorizeDevice": {
      "get": {
        "tags": [
          "Glimmr"
        ],
        "summary": "Trigger device authorization for the specified device.",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "The device Id to try authorizing.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/Glimmr/database": {
      "get": {
        "tags": [
          "Glimmr"
        ],
        "summary": "Download a backup of the current database.",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Glimmr"
        ],
        "summary": "Upload and replace the database with a copy from a db download.",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              },
              "encoding": {
                "files": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Glimmr/flashDevice": {
      "post": {
        "tags": [
          "Glimmr"
        ],
        "summary": "Flash an entire device.",
        "requestBody": {
          "description": "The ID of the device to flash on/off.",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "type": "string"
              }
            },
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Glimmr/flashLed": {
      "post": {
        "tags": [
          "Glimmr"
        ],
        "summary": "Flash a specific LED from the grid.",
        "requestBody": {
          "description": "The LED ID to flash.",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "application/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "text/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Glimmr/flashSector": {
      "post": {
        "tags": [
          "Glimmr"
        ],
        "summary": "Flash a specific Sector.",
        "requestBody": {
          "description": "The sector ID to flash.",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "application/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "text/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Glimmr/importAmbientScene": {
      "post": {
        "tags": [
          "Glimmr"
        ],
        "summary": "Upload and replace the database with a copy from a db download.",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              },
              "encoding": {
                "files": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Glimmr/importAudioScene": {
      "post": {
        "tags": [
          "Glimmr"
        ],
        "summary": "Upload and replace the database with a copy from a db download.",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              },
              "encoding": {
                "files": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Glimmr/startStream": {
      "post": {
        "tags": [
          "Glimmr"
        ],
        "summary": "Start Glimmr-to-Glimmr UDP stream",
        "requestBody": {
          "description": "A GlimmrData object containing the input dimensions of\r\nthe received colors.",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/GlimmrData"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GlimmrData"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GlimmrData"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GlimmrData"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AmbientScene": {
        "type": "object",
        "properties": {
          "AnimationTime": {
            "type": "number",
            "description": "Amount of time (in milliseconds) between color updates.",
            "format": "float"
          },
          "EasingTime": {
            "type": "number",
            "description": "How long to ease between colors.",
            "format": "float"
          },
          "Id": {
            "type": "integer",
            "description": "Theme ID. Will be auto-assigned for user defined themes.",
            "format": "int32"
          },
          "Easing": {
            "type": "string",
            "description": "Easing mode.\r\nBlend = 0 (Colors fade directly between one another)\r\nFadeIn = 1 (Color fades in after being replaced)\r\nFadeOut = 2 (Color fades out, is replaced, turns on full)\r\nFadeInOut = 3 (Colors fade in/out before being replaced)",
            "nullable": true
          },
          "Mode": {
            "type": "string",
            "description": "Animation mode.\r\nLinear = 0 (Colors progress normally)\r\nReverse = 1 (Colors progress in reverse direction)\r\nRandom = 2 (Colors are selected randomly)\r\nRandomAll = 3 (One random color for everything)\r\nLinearAll = 4 (Colors progress normally, one color for everything)",
            "nullable": true
          },
          "Name": {
            "type": "string",
            "description": "The theme name.",
            "nullable": true
          },
          "Colors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "An array of colors used by the scene.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}